Basic file upload with Angular-Django working
Got a piece of code on SF. In Angular used the input with type file to
upload the file in HTML. In the .ts file, created a File object which
takes this upload. The file is actually available as a list with one
item. Then a FormData is created and this file is appended to that and
passed to the backend as POST data. In the Django backend, this file
appears in request.FILES. To read the file access the object with the
key as specified in the FormData and use the .read() function. This
entire process works. Now to add UI and also bring in Pandas.