Created register, login and redirect
Registration is with the in-built Django User model. For logging-in a
separate UserToken model has been created like before. However, this
time I have a creation_time field that contains the time when the user
logs in. This will have to be updated to have a last_activity to store
the last value of time when the user who is logged in made a request on
the website. If no activity happens for a certain amount of time like a
few hours, the token can be deleted. So every time a request happens, a
check is needed to compare the current time with the last time of
activity.