Added user token clean up when session times out
This solves two problems. First, to deal with a session that has been
open for too long. When any request is made, it checks the difference
between the current time and the last updated time. Second, because this
is a token based authentication, if the user closes the window, the user
token stays in the database. A new window results in a new token. So
there can be several orphan tokens that have been created but never
deleted because the user never logged out. When a token has expired, it
will delete not only that token but all other tokens by the user so as
to clean them all up.