Repeat user login cleanup, file upload feature
When a user logins in again, there will be a check whether any previous
tokens of that user exist and if they do, they will be deleted. This
completes the user token handling. If a user leaves the session open for
too long, that token along with any other previous tokens will be
deleted. Only way orphan tokens can remain is if a user logins and
closes the window to never login again. Only way to deal with that is to
have a scheduled cleanup that checks all existing tokens in the
database. The file uploaded is received in the back-end. The only catch
was that it is received as a byte string and has to be decoded to UTF-8
as regular string. This can be directly written to a file. A media
directory has been created and called MEDIA_ROOT. Since, Django models
and model forms are not being used to upload, the upload to MEDIA_ROOT
is not automatic and the file write needs the MEDIA_ROOT specifically
joined to it.