Data file model, size checking and file upload UI
Created a data file model where every data file is linked to a user to
make sure files can be protected. A 2MB limit has been imposed on every
file upload as this will eventually be on the server. The file upload UI
has been improved a bit with an upload button and a success message when
the upload is successful. Next step is to create a list of files that a
user has uploaded and is available at every login. This list needs to
change on the fly as files are added and deleted.