Loaded file converted into Pandas object
The drop down list for choosing a file contains both the file name and
the user who has made the file available. This is particularly the case
with public files. With private files, the user should be implicit. The
backend will create a Pandas object every time a file is loaded. The
object remains in memory. Need to examine how this can be dealt with in
terms of memory.