Single file browser, file details edit form
Instead of two components for choosing user and public files, there need
only be one with a flag of whether it is for a user or not (True/False).
Also, on choosing a file, displays the file details. An edit option is
provided to be able to edit the details of the file.