Added analytics toolbar
Before loading the file was fecthing the entire file from the backend
and loading in component. This may not be a good idea for large files.
So now, loading will push the file into the list of loaded files. After
which the toolbar allows a user to query or plot. Both of these
functions are soon to come.