Query to fetch data frame from backend.
Now the load only fetches the head of the data frame and pushes the file
into a list. The query button on the toolbar fetches the data frame from
the backend. So far, clicking the query button will fetch the latest
file loaded. Next, to add a drop down list in the toolbar to choose
the file.